chore(data-access): move DotPageContentTypeService to data access lib
