+++ core-web/libs/data-access/src/index.ts
@@ -67,3 +67,4 @@ export * from './lib/dot-workflows-actions/dot-workflows-actions.service';
 export * from './lib/ema-app-configuration/ema-app-configuration.service';
 export * from './lib/paginator/paginator.service';
 export * from './lib/push-publish/push-publish.service';
+export * from './lib/dot-page-contenttype/dot-page-contenttype.service';
