+++ core-web/libs/dotcms-models/src/index.ts
@@ -84,3 +84,4 @@ export * from './lib/shared-models';
 export * from './lib/structure-type-view.model';
 export * from './lib/structure-type.model';
 export * from './lib/dot-folder.model';
+export * from './lib/dot-api-response';
